`CpCalculation`: add support for `AUTOPILOT` mode
The `CpParser` is also updated to add support for the new schema based
XML output format, which was already implemented for `pw.x` but is now
also used by `cp.x`.
