+++ aiida_quantumespresso/calculations/cp.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-"""Plugin to create a Quantum Espresso pw.x file.
+"""Plugin to create a Quantum Espresso cp.x file.
 
 TODO: COPY OUTDIR FROM PREVIOUS CALCULATION! Should be an input node of type
      RemoteData (or maybe subclass it?).
@@ -18,6 +18,7 @@ import os
 
 from aiida import orm
 from aiida.common.lang import classproperty
+from aiida.common import exceptions
 
 from aiida_quantumespresso.calculations import BasePwCpInputGenerator
 
@@ -29,10 +30,16 @@ class CpCalculation(BasePwCpInputGenerator):
     _CP_READ_UNIT_NUMBER = 50
     _CP_WRITE_UNIT_NUMBER = 51
     _FILE_XML_PRINT_COUNTER_BASENAME = 'print_counter.xml'
+    _FILE_PRINT_COUNTER_BASENAME = 'print_counter'
     _FILE_XML_PRINT_COUNTER = os.path.join(
         BasePwCpInputGenerator._OUTPUT_SUBFOLDER,
         f'{BasePwCpInputGenerator._PREFIX}_{_CP_WRITE_UNIT_NUMBER}.save',
         _FILE_XML_PRINT_COUNTER_BASENAME,
+    )
+    _FILE_PRINT_COUNTER = os.path.join(
+        BasePwCpInputGenerator._OUTPUT_SUBFOLDER,
+        f'{BasePwCpInputGenerator._PREFIX}_{_CP_WRITE_UNIT_NUMBER}.save',
+        _FILE_PRINT_COUNTER_BASENAME,
     )
 
     # Input file "sections" that we are going to write by calculation type
@@ -92,7 +99,7 @@ class CpCalculation(BasePwCpInputGenerator):
             BasePwCpInputGenerator._OUTPUT_SUBFOLDER,
             f'{BasePwCpInputGenerator._PREFIX}.{ext}',
         ) for ext in _cp_ext_list
-    ] + [_FILE_XML_PRINT_COUNTER]
+    ] + [_FILE_XML_PRINT_COUNTER, _FILE_PRINT_COUNTER]
 
     # in restarts, it will copy from the parent the following
     _restart_copy_from = os.path.join(
@@ -150,3 +157,37 @@ class CpCalculation(BasePwCpInputGenerator):
             message='The required POS file could not be read.')
         spec.exit_code(340, 'ERROR_READING_TRAJECTORY_DATA',
             message='The required trajectory data could not be read.')
+
+    @staticmethod
+    def _generate_PWCP_input_tail(*args, **kwargs):
+        """Parse CP specific input parameters."""
+        settings = kwargs['settings']
+
+        # AUTOPILOT
+        autopilot = settings.pop('AUTOPILOT', [])
+
+        if not autopilot:
+            return ''
+
+        autopilot_card = 'AUTOPILOT\n'
+
+        try:
+            for event in autopilot:
+                if isinstance(event['newvalue'], str):
+                    autopilot_card += f"ON_STEP = {event['onstep']} : '{event['what']}' = {event['newvalue']}\n"
+                else:
+                    autopilot_card += f"ON_STEP = {event['onstep']} : {event['what']} = {event['newvalue']}\n"
+        except KeyError as exception:
+            raise exceptions.InputValidationError(
+                f"""AUTOPILOT input: you must specify a list of dictionaries like the following:
+                 [
+                    {{'onstep' : 10, 'what' : 'dt', 'newvalue' : 5.0 }},
+                    {{'onstep' : 20, 'what' : 'whatever', 'newvalue' : 'pippo'}}
+                 ]
+                 You specified {autopilot}
+                 """
+            ) from exception
+
+        autopilot_card += 'ENDRULES\n'
+
+        return autopilot_card
