+++ aiida_quantumespresso/parsers/cp.py
@@ -5,8 +5,8 @@ import numpy
 from aiida.orm import Dict, TrajectoryData
 from qe_tools import CONSTANTS
 
-from aiida_quantumespresso.parsers.parse_raw.cp import parse_cp_raw_output, parse_cp_traj_stanzas
 from .base import Parser
+from .parse_raw.cp import parse_cp_raw_output, parse_cp_traj_stanzas
 
 
 class CpParser(Parser):
@@ -35,183 +35,219 @@ class CpParser(Parser):
         elif len(xml_files) > 1:
             return self.exit(self.exit_codes.ERROR_OUTPUT_XML_MULTIPLE)
 
-        if self.node.process_class._FILE_XML_PRINT_COUNTER_BASENAME not in list_of_files:
-            self.logger.error('We could not find the print counter file in the output')
-            # TODO: Add an error for this counter
-            return self.exit(self.exit_codes.ERROR_MISSING_XML_FILE)
+        # cp.x can produce, depending on the particular version of the code, a file called `print_counter.xml` or
+        # `print_counter`, which is a plain text file with the number of the last timestep written in the trajectory
+        # output. Note that if no trajectory is produced (for example because a single conjugate gradient step was
+        # performed to calculate the ground state and the wavefunctions velocities) no printer_counter* file is written.
+
+        print_counter_xml = True
+        no_trajectory_output = False
+
+        filename_counter_txt = self.node.process_class._FILE_PRINT_COUNTER_BASENAME
+        filename_counter_xml = self.node.process_class._FILE_XML_PRINT_COUNTER_BASENAME
+
+        # The following can happen and is not an error!
+        if filename_counter_xml not in list_of_files and filename_counter_txt not in list_of_files:
+            self.logger.error(
+                f'We could not find the print counter file (`{filename_counter_txt}` or `{filename_counter_xml}`), '
+                'assuming no trajectory output was produced'
+            )
+            no_trajectory_output = True
+
+        if not no_trajectory_output:
+            if filename_counter_txt in list_of_files:
+                self.logger.info('print counter not in xml format')
+                print_counter_xml = False
+                filename_counter = filename_counter_txt
+            else:  # xml format
+                print_counter_xml = True
+                self.logger.info('print counter in xml format')
+                filename_counter = filename_counter_xml
 
         output_stdout = retrieved.get_object_content(stdout_filename)
         output_xml = retrieved.get_object_content(xml_files[0])
-        output_xml_counter = retrieved.get_object_content(self.node.process_class._FILE_XML_PRINT_COUNTER_BASENAME)
-        out_dict, _raw_successful = parse_cp_raw_output(output_stdout, output_xml, output_xml_counter)
-
-        # parse the trajectory. Units in Angstrom, picoseconds and eV.
-        # append everthing in the temporary dictionary raw_trajectory
-        raw_trajectory = {}
-        evp_keys = [
-            'electronic_kinetic_energy', 'cell_temperature', 'ionic_temperature', 'scf_total_energy', 'enthalpy',
-            'enthalpy_plus_kinetic', 'energy_constant_motion', 'volume', 'pressure'
-        ]
-
-        # Now prepare the reordering, as filex in the xml are  ordered
-        reordering = self._generate_sites_ordering(out_dict['species'], out_dict['atoms'])
-
-        pos_filename = f'{self.node.process_class._PREFIX}.pos'
-        if pos_filename not in list_of_files:
-            return self.exit(self.exit_codes.ERROR_READING_POS_FILE)
-
-        trajectories = [
-            ('positions', 'pos', CONSTANTS.bohr_to_ang, out_dict['number_of_atoms']),
-            ('cells', 'cel', CONSTANTS.bohr_to_ang, 3),
-            (
-                'velocities', 'vel', CONSTANTS.bohr_to_ang / CONSTANTS.timeau_to_sec * 10**12,
-                out_dict['number_of_atoms']
-            ),
-        ]
-
-        for name, extension, scale, elements in trajectories:
-            try:
-                with retrieved.open(f'{self.node.process_class._PREFIX}.{extension}') as datafile:
-                    data = [l.split() for l in datafile]
-                    # POSITIONS stored in angstrom
-                traj_data = parse_cp_traj_stanzas(
-                    num_elements=elements, splitlines=data, prepend_name=f'{name}_traj', rescale=scale
-                )
-                # here initialize the dictionary. If the parsing of positions fails, though, I don't have anything
-                # out of the CP dynamics. Therefore, the calculation status is set to FAILED.
-                if extension != 'cel':
-                    raw_trajectory[f'{name}_ordered'] = self._get_reordered_array(
-                        traj_data[f'{name}_traj_data'], reordering
-                    )
-                else:
-                    raw_trajectory['cells'] = numpy.array(traj_data['cells_traj_data'])
-                if extension == 'pos':
-                    raw_trajectory['times'] = numpy.array(traj_data[f'{name}_traj_times'])
-            except IOError:
-                out_dict['warnings'].append(f'Unable to open the {extension.upper()} file... skipping.')
-
-        # =============== EVP trajectory ============================
-        try:
-            with retrieved.open(f'{self._node.process_class._PREFIX}.evp') as handle:
-                matrix = numpy.genfromtxt(handle)
-            # there might be a different format if the matrix has one row only
-            try:
-                matrix.shape[1]
-            except IndexError:
-                matrix = numpy.array(numpy.matrix(matrix))
-
-            if LooseVersion(out_dict['creator_version']) > LooseVersion('5.1'):
-                # Between version 5.1 and 5.1.1, someone decided to change
-                # the .evp output format, without any way to know that this
-                # happened... SVN commit 11158.
-                # I here use the version number to parse, plus some
-                # heuristics to check that I'm doing the right thing
-                #print "New version"
-                raw_trajectory['steps'] = numpy.array(matrix[:, 0], dtype=int)
-                raw_trajectory['evp_times'] = matrix[:, 1]  # TPS, ps
-                raw_trajectory['electronic_kinetic_energy'] = matrix[:, 2] * CONSTANTS.hartree_to_ev  # EKINC, eV
-                raw_trajectory['cell_temperature'] = matrix[:, 3]  # TEMPH, K
-                raw_trajectory['ionic_temperature'] = matrix[:, 4]  # TEMPP, K
-                raw_trajectory['scf_total_energy'] = matrix[:, 5] * CONSTANTS.hartree_to_ev  # ETOT, eV
-                raw_trajectory['enthalpy'] = matrix[:, 6] * CONSTANTS.hartree_to_ev  # ENTHAL, eV
-                raw_trajectory['enthalpy_plus_kinetic'] = matrix[:, 7] * CONSTANTS.hartree_to_ev  # ECONS, eV
-                raw_trajectory['energy_constant_motion'] = matrix[:, 8] * CONSTANTS.hartree_to_ev  # ECONT, eV
-                raw_trajectory['volume'] = matrix[:, 9] * (CONSTANTS.bohr_to_ang**3)  # volume, angstrom^3
-                raw_trajectory['pressure'] = matrix[:, 10]  # out_press, GPa
-            else:
-                #print "Old version"
-                raw_trajectory['steps'] = numpy.array(matrix[:, 0], dtype=int)
-                raw_trajectory['electronic_kinetic_energy'] = matrix[:, 1] * CONSTANTS.hartree_to_ev  # EKINC, eV
-                raw_trajectory['cell_temperature'] = matrix[:, 2]  # TEMPH, K
-                raw_trajectory['ionic_temperature'] = matrix[:, 3]  # TEMPP, K
-                raw_trajectory['scf_total_energy'] = matrix[:, 4] * CONSTANTS.hartree_to_ev  # ETOT, eV
-                raw_trajectory['enthalpy'] = matrix[:, 5] * CONSTANTS.hartree_to_ev  # ENTHAL, eV
-                raw_trajectory['enthalpy_plus_kinetic'] = matrix[:, 6] * CONSTANTS.hartree_to_ev  # ECONS, eV
-                raw_trajectory['energy_constant_motion'] = matrix[:, 7] * CONSTANTS.hartree_to_ev  # ECONT, eV
-                raw_trajectory['volume'] = matrix[:, 8] * (CONSTANTS.bohr_to_ang**3)  # volume, angstrom^3
-                raw_trajectory['pressure'] = matrix[:, 9]  # out_press, GPa
-                raw_trajectory['evp_times'] = matrix[:, 10]  # TPS, ps
-
-            # Huristics to understand if it's correct.
-            # A better heuristics could also try to fix possible issues
-            # (in new versions of QE, it's possible to recompile it with
-            # the __OLD_FORMAT flag to get back the old version format...)
-            # but I won't do it, as there may be also other columns swapped.
-            # Better to stop and ask the user to check what's going on.
-            max_time_difference = abs(numpy.array(raw_trajectory['times']) -
-                                      numpy.array(raw_trajectory['evp_times'])).max()
-            if max_time_difference > 1.e-4:  # It is typically ~1.e-7 due to roundoff errors
-                # If there is a large discrepancy
-                # it means there is something very weird going on...
-                return self.exit(self.exit_codes.ERROR_READING_TRAJECTORY_DATA)
-
-            # Delete evp_times in any case, it's a duplicate of 'times'
-            del raw_trajectory['evp_times']
-        except IOError:
-            out_dict['warnings'].append('Unable to open the EVP file... skipping.')
-
-        # get the symbols from the input
-        # TODO: I should have kinds in TrajectoryData
-        input_structure = self.node.inputs.structure
-        raw_trajectory['symbols'] = [str(i.kind_name) for i in input_structure.sites]
-
-        traj = TrajectoryData()
-        traj.set_trajectory(
-            stepids=raw_trajectory['steps'],
-            cells=raw_trajectory['cells'],
-            symbols=raw_trajectory['symbols'],
-            positions=raw_trajectory['positions_ordered'],
-            times=raw_trajectory['times'],
-            velocities=raw_trajectory['velocities_ordered'],
+        output_xml_counter = None if no_trajectory_output else retrieved.get_object_content(filename_counter)
+        out_dict, _raw_successful = parse_cp_raw_output(
+            output_stdout, output_xml, output_xml_counter, print_counter_xml
         )
 
-        for this_name in evp_keys:
-            try:
-                traj.set_array(this_name, raw_trajectory[this_name])
-            except KeyError:
-                # Some columns may have not been parsed, skip
-                pass
+        if not no_trajectory_output:
+            # parse the trajectory. Units in Angstrom, picoseconds and eV.
+            # append everthing in the temporary dictionary raw_trajectory
+            raw_trajectory = {}
+            evp_keys = [
+                'electronic_kinetic_energy', 'cell_temperature', 'ionic_temperature', 'scf_total_energy', 'enthalpy',
+                'enthalpy_plus_kinetic', 'energy_constant_motion', 'volume', 'pressure'
+            ]
 
-        self.out('output_trajectory', traj)
+            # order of atom in the output trajectory changed somewhere after 6.5
+            if LooseVersion(out_dict['creator_version']) > LooseVersion('6.5'):
+                new_cp_ordering = True
+            else:
+                new_cp_ordering = False
+
+            # Now prepare the reordering, as files in the xml are ordered
+            if new_cp_ordering:
+                reordering = None
+            else:
+                try:
+                    # this works for old xml only
+                    reordering = self._generate_sites_ordering(out_dict['species'], out_dict['atoms'])
+                except KeyError:
+                    # this works for newer versions
+                    reordering = self._generate_sites_ordering(
+                        out_dict['structure']['species'], out_dict['structure']['atoms']
+                    )
+
+            pos_filename = f'{self.node.process_class._PREFIX}.pos'
+            if pos_filename not in list_of_files:
+                out_dict['warnings'].append('Unable to open the POS file... skipping.')
+                return self.exit_codes.ERROR_READING_POS_FILE
+            number_of_atoms = out_dict.get(
+                'number_of_atoms', out_dict['structure']['number_of_atoms'] if 'structure' in out_dict else None
+            )
+            trajectories = [
+                ('positions', 'pos', CONSTANTS.bohr_to_ang, number_of_atoms),
+                ('cells', 'cel', CONSTANTS.bohr_to_ang, 3),
+                ('velocities', 'vel', CONSTANTS.bohr_to_ang / (CONSTANTS.timeau_to_sec * 10**12), number_of_atoms),
+                ('forces', 'for', CONSTANTS.hartree_to_ev / CONSTANTS.bohr_to_ang, number_of_atoms),
+            ]
+
+            for name, extension, scale, elements in trajectories:
+                try:
+                    with retrieved.open(f'{self.node.process_class._PREFIX}.{extension}') as datafile:
+                        data = [l.split() for l in datafile]
+                        # POSITIONS stored in angstrom
+                    traj_data = parse_cp_traj_stanzas(
+                        num_elements=elements, splitlines=data, prepend_name=f'{name}_traj', rescale=scale
+                    )
+                    # here initialize the dictionary. If the parsing of positions fails, though, I don't have anything
+                    # out of the CP dynamics. Therefore, the calculation status is set to FAILED.
+                    if extension != 'cel':
+                        raw_trajectory[f'{name}_ordered'] = self._get_reordered_array(
+                            traj_data[f'{name}_traj_data'], reordering
+                        )
+                    else:
+                        # NOTE: the trajectory output has the cell matrix transposed!!
+                        raw_trajectory['cells'] = numpy.array(traj_data['cells_traj_data']).transpose((0, 2, 1))
+                    if extension == 'pos':
+                        raw_trajectory['traj_times'] = numpy.array(traj_data[f'{name}_traj_times'])
+                except IOError:
+                    out_dict['warnings'].append(f'Unable to open the {extension.upper()} file... skipping.')
+
+            # =============== EVP trajectory ============================
+            try:
+                with retrieved.open(f'{self._node.process_class._PREFIX}.evp') as handle:
+                    matrix = numpy.genfromtxt(handle)
+                # there might be a different format if the matrix has one row only
+                try:
+                    matrix.shape[1]
+                except IndexError:
+                    matrix = numpy.array(numpy.matrix(matrix))
+
+                if LooseVersion(out_dict['creator_version']) > LooseVersion('5.1'):
+                    # Between version 5.1 and 5.1.1, someone decided to change
+                    # the .evp output format, without any way to know that this
+                    # happened... SVN commit 11158.
+                    # I here use the version number to parse, plus some
+                    # heuristics to check that I'm doing the right thing
+                    #print "New version"
+                    raw_trajectory['steps'] = numpy.array(matrix[:, 0], dtype=int)
+                    raw_trajectory['times'] = matrix[:, 1]  # TPS, ps
+                    raw_trajectory['electronic_kinetic_energy'] = matrix[:, 2] * CONSTANTS.hartree_to_ev  # EKINC, eV
+                    raw_trajectory['cell_temperature'] = matrix[:, 3]  # TEMPH, K
+                    raw_trajectory['ionic_temperature'] = matrix[:, 4]  # TEMPP, K
+                    raw_trajectory['scf_total_energy'] = matrix[:, 5] * CONSTANTS.hartree_to_ev  # ETOT, eV
+                    raw_trajectory['enthalpy'] = matrix[:, 6] * CONSTANTS.hartree_to_ev  # ENTHAL, eV
+                    raw_trajectory['enthalpy_plus_kinetic'] = matrix[:, 7] * CONSTANTS.hartree_to_ev  # ECONS, eV
+                    raw_trajectory['energy_constant_motion'] = matrix[:, 8] * CONSTANTS.hartree_to_ev  # ECONT, eV
+                    raw_trajectory['volume'] = matrix[:, 9] * (CONSTANTS.bohr_to_ang**3)  # volume, angstrom^3
+                    raw_trajectory['pressure'] = matrix[:, 10]  # out_press, GPa
+                else:
+                    #print "Old version"
+                    raw_trajectory['steps'] = numpy.array(matrix[:, 0], dtype=int)
+                    raw_trajectory['electronic_kinetic_energy'] = matrix[:, 1] * CONSTANTS.hartree_to_ev  # EKINC, eV
+                    raw_trajectory['cell_temperature'] = matrix[:, 2]  # TEMPH, K
+                    raw_trajectory['ionic_temperature'] = matrix[:, 3]  # TEMPP, K
+                    raw_trajectory['scf_total_energy'] = matrix[:, 4] * CONSTANTS.hartree_to_ev  # ETOT, eV
+                    raw_trajectory['enthalpy'] = matrix[:, 5] * CONSTANTS.hartree_to_ev  # ENTHAL, eV
+                    raw_trajectory['enthalpy_plus_kinetic'] = matrix[:, 6] * CONSTANTS.hartree_to_ev  # ECONS, eV
+                    raw_trajectory['energy_constant_motion'] = matrix[:, 7] * CONSTANTS.hartree_to_ev  # ECONT, eV
+                    raw_trajectory['volume'] = matrix[:, 8] * (CONSTANTS.bohr_to_ang**3)  # volume, angstrom^3
+                    raw_trajectory['pressure'] = matrix[:, 9]  # out_press, GPa
+                    raw_trajectory['times'] = matrix[:, 10]  # TPS, ps
+
+                # Huristics to understand if it's correct.
+                # A better heuristics could also try to fix possible issues
+                # (in new versions of QE, it's possible to recompile it with
+                # the __OLD_FORMAT flag to get back the old version format...)
+                # but I won't do it, as there may be also other columns swapped.
+                # Better to stop and ask the user to check what's going on.
+
+                #work around for 100ps format bug
+                mask = numpy.array(raw_trajectory['traj_times']) >= 0
+                len_bugged = len(numpy.array(raw_trajectory['times'])[mask == False])
+                len_ok = len(numpy.array(raw_trajectory['times'])[mask])
+                if len_ok > 0:
+                    max_time_difference = abs(
+                        numpy.array(raw_trajectory['times'])[mask] - numpy.array(raw_trajectory['traj_times'])[mask]
+                    ).max()
+                else:
+                    max_time_difference = 0.0
+
+                if max_time_difference > 1.e-4 or (
+                    len_bugged > 0 and numpy.array(raw_trajectory['times'])[mask == False].min() < 100.0
+                ):  # It is typically ~1.e-7 due to roundoff errors
+                    # If there is a large discrepancy
+                    # it means there is something very weird going on...
+                    return self.exit_codes.ERROR_READING_TRAJECTORY_DATA
+
+                # keep both times array (that usually are duplicated)
+                # so that the user can check them by himselves
+                if len_bugged > 0:
+                    out_dict['warnings'].append(
+                        '100ps format bug detected: ignoring trajectory\'s printed time from 100ps on'
+                    )
+            except IOError:
+                out_dict['warnings'].append('Unable to open the EVP file... skipping.')
+
+            # get the symbols from the input
+            # TODO: I should have kinds in TrajectoryData
+            input_structure = self.node.inputs.structure
+            raw_trajectory['symbols'] = [str(i.kind_name) for i in input_structure.sites]
+
+            traj = TrajectoryData()
+            traj.set_trajectory(
+                stepids=raw_trajectory['steps'],
+                cells=raw_trajectory['cells'],
+                symbols=raw_trajectory['symbols'],
+                positions=raw_trajectory['positions_ordered'],
+                times=raw_trajectory['times'],
+                velocities=raw_trajectory['velocities_ordered'],
+            )
+
+            # eventually set the forces
+            try:
+                traj.set_array('forces', raw_trajectory['forces_ordered'])
+            except KeyError:
+                out_dict['warnings'].append('failed to set forces')
+
+            for this_name in evp_keys:
+                try:
+                    traj.set_array(this_name, raw_trajectory[this_name])
+                except KeyError:
+                    # Some columns may have not been parsed, skip
+                    pass
+
+            self.out('output_trajectory', traj)
 
         # Remove big dictionaries that would be redundant
-        # For atoms and cell, there is a small possibility that nothing is parsed
-        # but then probably nothing moved.
-        try:
-            del out_dict['atoms']
-        except KeyError:
-            pass
-        try:
-            del out_dict['cell']
-        except KeyError:
-            pass
-        try:
-            del out_dict['ions_positions_stau']
-        except KeyError:
-            pass
-        try:
-            del out_dict['ions_positions_svel']
-        except KeyError:
-            pass
-        try:
-            del out_dict['ions_positions_taui']
-        except KeyError:
-            pass
-        # This should not be needed
-        try:
-            del out_dict['atoms_index_list']
-        except KeyError:
-            pass
-        # This should be already in the input
-        try:
-            del out_dict['atoms_if_pos_list']
-        except KeyError:
-            pass
-        #
-        try:
-            del out_dict['ions_positions_force']
-        except KeyError:
-            pass
+        # For atoms and cell, there is a small possibility that nothing is parsed but then probably nothing moved.
+        for key in [
+            'atoms', 'cell', 'ions_positions_stau', 'ions_positions_svel', 'ions_positions_taui', 'atoms_index_list',
+            'atoms_if_pos_list', 'ions_positions_force', 'bands', 'structure'
+        ]:
+            out_dict.pop(key, None)
 
         # convert the dictionary into an AiiDA object
         output_params = Dict(dict=out_dict)
@@ -267,4 +303,7 @@ class CpParser(Parser):
         return [origlist[e] for e in reordering]
 
     def _get_reordered_array(self, _input, reordering):
-        return numpy.array([self._get_reordered_list(i, reordering) for i in _input])
+        if reordering is not None:
+            return numpy.array([self._get_reordered_list(i, reordering) for i in _input])
+        else:
+            return numpy.array(_input)
