+++ tests/conftest.py
@@ -237,16 +237,23 @@ def generate_upf_family(tmp_path, generate_upf_data):
 def generate_structure():
     """Return a `StructureData` representing bulk silicon."""
 
-    def _generate_structure():
-        """Return a `StructureData` representing bulk silicon."""
+    def _generate_structure(structure_id='silicon'):
+        """Return a `StructureData` representing bulk silicon or a snapshot of a single water molecule dynamics."""
         from aiida.orm import StructureData
 
-        param = 5.43
-        cell = [[param / 2., param / 2., 0], [param / 2., 0, param / 2.], [0, param / 2., param / 2.]]
-        structure = StructureData(cell=cell)
-        structure.append_atom(position=(0., 0., 0.), symbols='Si', name='Si')
-        structure.append_atom(position=(param / 4., param / 4., param / 4.), symbols='Si', name='Si')
-
+        if structure_id == 'silicon':
+            param = 5.43
+            cell = [[param / 2., param / 2., 0], [param / 2., 0, param / 2.], [0, param / 2., param / 2.]]
+            structure = StructureData(cell=cell)
+            structure.append_atom(position=(0., 0., 0.), symbols='Si', name='Si')
+            structure.append_atom(position=(param / 4., param / 4., param / 4.), symbols='Si', name='Si')
+        elif structure_id == 'water':
+            structure = StructureData(cell=[[5.29177209, 0., 0.], [0., 5.29177209, 0.], [0., 0., 5.29177209]])
+            structure.append_atom(position=[12.73464656, 16.7741411, 24.35076238], symbols='H', name='H')
+            structure.append_atom(position=[-29.3865565, 9.51707929, -4.02515904], symbols='H', name='H')
+            structure.append_atom(position=[1.04074437, -1.64320127, -1.27035021], symbols='O', name='O')
+        else:
+            raise KeyError('Unknown structure_id=\'{}\''.format(structure_id))
         return structure
 
     return _generate_structure
@@ -442,3 +449,51 @@ def generate_inputs_pw(fixture_code, generate_structure, generate_kpoints_mesh, 
         return inputs
 
     return _generate_inputs_pw
+
+
+@pytest.fixture
+def generate_inputs_cp(fixture_code, generate_structure, generate_upf_data):
+    """Generate default inputs for a CpCalculation."""
+
+    def _generate_inputs_cp(autopilot=False):
+        """Generate default inputs for a CpCalculation."""
+        from aiida.orm import Dict
+        from aiida_quantumespresso.utils.resources import get_default_options
+
+        inputs = {
+            'code': fixture_code('quantumespresso.cp'),
+            'structure': generate_structure(),
+            'parameters': Dict(dict={
+                'CONTROL': {
+                    'calculation': 'cp'
+                },
+                'SYSTEM': {
+                    'ecutrho': 240.0,
+                    'ecutwfc': 30.0
+                }
+            }),
+            'pseudos': {
+                'Si': generate_upf_data('Si')
+            },
+            'metadata': {
+                'options': get_default_options()
+            }
+        }
+        if autopilot:
+            inputs['settings'] = Dict(
+                dict={
+                    'AUTOPILOT': [{
+                        'onstep': 2,
+                        'what': 'dt',
+                        'newvalue': 42.0
+                    }, {
+                        'onstep': 3,
+                        'what': 'dt',
+                        'newvalue': 42.42
+                    }]
+                }
+            )
+
+        return inputs
+
+    return _generate_inputs_cp
