+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.0_0
@@ -1,0 +1,18 @@
+           1  -17.174379109009372        0.0000000000000000        0.0000000000000000
+CG1      -17.1743791090      -17.1743839820      -17.1743840284      -17.1743840283
+CG2 1.0000000 0.3000000 0.2734494 0.4999882
+           2  -17.174384028253794        0.0000000000000000        0.0000000000000000
+CG1      -17.1743840283      -17.1743841671      -17.1743843595      -17.1743843594
+CG2 1.0000000 0.5468989 0.3103656 0.4998717
+           3  -17.174384359393361        0.0000000000000000        0.0000000000000000
+CG1      -17.1743843594      -17.1743844164      -17.1743844324      -17.1743844327
+CG2 1.0000000 0.6207311 0.4227147 0.5021966
+           4  -17.174384432687461        0.0000000000000000        0.0000000000000000
+CG1      -17.1743844327      -17.1743844299      -17.1743844623      -17.1743844635
+CG2 1.0000000 0.8454295 0.4132247 0.5196914
+           5  -17.174384463458303        0.0000000000000000        0.0000000000000000
+CG1      -17.1743844635      -17.1743844561      -17.1743844711      -17.1743844713
+CG2 1.0000000 0.8264495 0.3442924 0.5118901
+           6  -17.174384471275090        0.0000000000000000        0.0000000000000000
+CG1      -17.1743844713      -17.1743844720      -17.1743844735      -17.1743844732
+CG2 1.0000000 0.6885848 0.3773207 0.4332206
