+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.cel
@@ -1,0 +1,32 @@
+     10  0.00224956
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     12  0.00239470
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     15  0.00261240
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     18  0.00283009
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     21  0.00304779
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     24  0.00326549
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     27  0.00348319
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     30  0.00370089
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
