+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.eig
@@ -1,0 +1,24 @@
+  STEP:     10       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.49  -12.73   -8.23   -6.42
+  STEP:     12       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.53  -12.77   -8.26   -6.47
+  STEP:     15       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.37  -12.67   -8.16   -6.39
+  STEP:     18       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.32  -12.66   -8.12   -6.38
+  STEP:     21       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.33  -12.70   -8.13   -6.39
+  STEP:     24       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.20  -12.63   -8.08   -6.34
+  STEP:     27       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.20  -12.60   -8.08   -6.36
+  STEP:     30       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.15  -12.57   -8.05   -6.35
