+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.evp
@@ -1,0 +1,8 @@
+     10  2.249562E-03  6.361068E-04  0.000000E+00  2.909040E+02        -17.17438447        -17.17438447        -17.17023890        -17.16960279  1.000000E+03      -0.00000
+     12  2.394695E-03  2.021100E-04  0.000000E+00  3.076704E+02        -17.17412415        -17.17412415        -17.16973964        -17.16953753  1.000000E+03      -0.00000
+     15  2.612395E-03  7.835814E-05  0.000000E+00  3.334343E+02        -17.17442201        -17.17442201        -17.16967035        -17.16959199  1.000000E+03      -0.00000
+     18  2.830095E-03  3.436010E-04  0.000000E+00  3.103230E+02        -17.17433368        -17.17433368        -17.16991137        -17.16956777  1.000000E+03      -0.00000
+     21  3.047794E-03  2.497460E-04  0.000000E+00  3.116892E+02        -17.17424329        -17.17424329        -17.16980152        -17.16955177  1.000000E+03      -0.00000
+     24  3.265494E-03  1.373922E-04  0.000000E+00  3.212534E+02        -17.17430130        -17.17430130        -17.16972323        -17.16958583  1.000000E+03      -0.00000
+     27  3.483193E-03  2.988302E-04  0.000000E+00  2.933242E+02        -17.17404099        -17.17404099        -17.16986092        -17.16956209  1.000000E+03      -0.00000
+     30  3.700893E-03  2.821960E-04  0.000000E+00  2.797270E+02        -17.17382794        -17.17382794        -17.16984165        -17.16955945  1.000000E+03      -0.00000
