+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.nos
@@ -1,0 +1,8 @@
+     10  2.249562E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     12  2.394695E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     15  2.612395E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     18  2.830095E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     21  3.047794E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     24  3.265494E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     27  3.483193E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     30  3.700893E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
