+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.pos
@@ -1,0 +1,32 @@
+     10  0.00224956
+     0.55092489705446E+00     0.89015870064501E+00     0.10589905087984E+01
+    -0.38553733104569E+00    -0.43136591605268E+00    -0.14265117462520E+01
+    -0.31806693537555E+00    -0.50542901591202E+00     0.41024598359085E+00
+     12  0.00239470
+     0.55272147991935E+00     0.89175857893172E+00     0.10665465827273E+01
+    -0.39465687794958E+00    -0.43040754600185E+00    -0.14304085998814E+01
+    -0.31760925012311E+00    -0.50558890643312E+00     0.41001728232213E+00
+     15  0.00261240
+     0.55613611707251E+00     0.89534221684383E+00     0.10784454410845E+01
+    -0.40871477308902E+00    -0.42861229794304E+00    -0.14362163912582E+01
+    -0.31694404649897E+00    -0.50592508680630E+00     0.40963659063586E+00
+     18  0.00283009
+     0.56014821010685E+00     0.89997817283135E+00     0.10902564612997E+01
+    -0.42246533537203E+00    -0.42646027342670E+00    -0.14414756310823E+01
+    -0.31633539217093E+00    -0.50634933558779E+00     0.40922710436142E+00
+     21  0.00304779
+     0.56466212510982E+00     0.90542866052214E+00     0.11016191723925E+01
+    -0.43553966135344E+00    -0.42400192772153E+00    -0.14459797847172E+01
+    -0.31580036648477E+00    -0.50684363767504E+00     0.40879844452030E+00
+     24  0.00326549
+     0.56962986127207E+00     0.91162427433000E+00     0.11130802418019E+01
+    -0.44872716085447E+00    -0.42111528462533E+00    -0.14499465643365E+01
+    -0.31528663127610E+00    -0.50741127873154E+00     0.40833005140841E+00
+     27  0.00348319
+     0.57472227351194E+00     0.91824237617228E+00     0.11241032654022E+01
+    -0.46160310287754E+00    -0.41780277027431E+00    -0.14530007428545E+01
+    -0.31480016066465E+00    -0.50803194224362E+00     0.40783199859077E+00
+     30  0.00370089
+     0.57989297408593E+00     0.92507217560154E+00     0.11343356863191E+01
+    -0.47379308119608E+00    -0.41410543742723E+00    -0.14550518349346E+01
+    -0.31436145580562E+00    -0.50868988801089E+00     0.40732066553847E+00
