+++ tests/parsers/fixtures/cp/6.6_autopilot/aiida.vel
@@ -1,0 +1,32 @@
+     10  0.00224956
+     0.26693656288479E-03     0.21352115337272E-03     0.12443764458129E-02
+    -0.15179329302867E-02     0.14680867011606E-03    -0.65556606812030E-03
+     0.78187272962623E-04    -0.22520613968043E-04    -0.36800648605800E-04
+     12  0.00239470
+     0.33313251660290E-03     0.32140406135440E-03     0.12855157072344E-02
+    -0.15363128568950E-02     0.17511611093835E-03    -0.64927985563230E-03
+     0.75198771268258E-04    -0.31032510768303E-04    -0.39764740725132E-04
+     15  0.00261240
+     0.41774170427291E-03     0.46447567865315E-03     0.13365556262702E-02
+    -0.15657747278498E-02     0.22178504936065E-03    -0.62653369323156E-03
+     0.71752063973557E-04    -0.42891295500870E-04    -0.44376370814912E-04
+     18  0.00283009
+     0.47217108088044E-03     0.56046489095185E-03     0.12784267006172E-02
+    -0.14790811195376E-02     0.25482300205279E-03    -0.53728706563079E-03
+     0.62931877416056E-04    -0.50955493312796E-04    -0.46321227186660E-04
+     21  0.00304779
+     0.53103618659804E-03     0.65056816491466E-03     0.12655668801267E-02
+    -0.14519765029592E-02     0.29519972868023E-03    -0.47200395959790E-03
+     0.57558769772578E-04    -0.59110493349686E-04    -0.49597682533047E-04
+     24  0.00326549
+     0.56348932993541E-03     0.71756212237088E-03     0.12647059676222E-02
+    -0.14644775902147E-02     0.34587493254776E-03    -0.39783107405493E-03
+     0.56311766267458E-04    -0.66464815932423E-04    -0.54179680847958E-04
+     27  0.00348319
+     0.56796481566599E-03     0.74866813677486E-03     0.11752771772596E-02
+    -0.13856036166382E-02     0.38885800205220E-03    -0.27904250502619E-03
+     0.51102425060757E-04    -0.71095383676681E-04    -0.56014667014585E-04
+     30  0.00370089
+     0.58210437808501E-03     0.76675084385057E-03     0.11130803928384E-02
+    -0.13432679953506E-02     0.43502844707613E-03    -0.18399118317460E-03
+     0.47572726079107E-04    -0.75111205682920E-04    -0.58068075603987E-04
