+++ tests/parsers/fixtures/cp/6.6_cgstep/aiida.0_0
@@ -1,0 +1,27 @@
+           1  -17.175036121137996        0.0000000000000000        0.0000000000000000
+CG1      -17.1750361211      -17.1753838057      -17.1753856202      -17.1753856152
+CG2 1.0000000 0.3000000 0.2798367 0.4999929
+           2  -17.175385615242106        0.0000000000000000        0.0000000000000000
+CG1      -17.1753856152      -17.1753909391      -17.1754069261      -17.1754069270
+CG2 1.0000000 0.5596734 0.2999111 0.5000233
+           3  -17.175406927044374        0.0000000000000000        0.0000000000000000
+CG1      -17.1754069270      -17.1754093582      -17.1754106340      -17.1754106328
+CG2 1.0000000 0.5998222 0.3780396 0.4998328
+           4  -17.175410632799824        0.0000000000000000        0.0000000000000000
+CG1      -17.1754106328      -17.1754107630      -17.1754118328      -17.1754118276
+CG2 1.0000000 0.7560792 0.3888914 0.4978039
+           5  -17.175411827559454        0.0000000000000000        0.0000000000000000
+CG1      -17.1754118276      -17.1754117660      -17.1754121857      -17.1754121854
+CG2 1.0000000 0.7777829 0.3734859 0.4996221
+           6  -17.175412185428897        0.0000000000000000        0.0000000000000000
+CG1      -17.1754121854      -17.1754121837      -17.1754122857      -17.1754122863
+CG2 1.0000000 0.7469718 0.3719198 0.5031916
+           7  -17.175412286320931        0.0000000000000000        0.0000000000000000
+CG1      -17.1754122863      -17.1754122984      -17.1754123217      -17.1754123219
+CG2 1.0000000 0.7438396 0.4107888 0.5038625
+           8  -17.175412321923879        0.0000000000000000        0.0000000000000000
+CG1      -17.1754123219      -17.1754123198      -17.1754123346      -17.1754123354
+CG2 1.0000000 0.8215776 0.3949041 0.5307624
+           9  -17.175412335408168        0.0000000000000000        0.0000000000000000
+CG1      -17.1754123354      -17.1754123356      -17.1754123393      -17.1754123389
+CG2 1.0000000 0.7898081 0.4004585 0.4550937
