+++ tests/parsers/fixtures/cp/6.6_cgsteps/aiida.0_0
@@ -1,0 +1,15 @@
+           1  -17.174317073266440        0.0000000000000000        0.0000000000000000
+CG1      -17.1743170733      -17.1743171105      -17.1743171108      -17.1743171108
+CG2 1.0000000 0.3000000 0.2760087 0.4999998
+           2  -17.174317110824859        0.0000000000000000        0.0000000000000000
+CG1      -17.1743171108      -17.1743171107      -17.1743171129      -17.1743171129
+CG2 1.0000000 0.5520173 0.2731760 0.5000089
+           3  -17.174317112921550        0.0000000000000000        0.0000000000000000
+CG1      -17.1743171129      -17.1743171131      -17.1743171132      -17.1743171132
+CG2 1.0000000 0.5463520 0.3366769 0.4998841
+           4  -17.174317113200065        0.0000000000000000        0.0000000000000000
+CG1      -17.1743171132      -17.1743171133      -17.1743171133      -17.1743171134
+CG2 1.0000000 0.6733538 0.5248939 0.5720589
+           5  -17.174317113355755        0.0000000000000000        0.0000000000000000
+CG1      -17.1743171134      -17.1743171133      -17.1743171134      -17.1743171136
+CG2 1.0000000 1.0497879 0.4453073 1.7346468
