+++ tests/parsers/fixtures/cp/6.6_cgsteps/aiida.cel
@@ -1,0 +1,12 @@
+     10  0.00072567
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     20  0.00145133
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
+     30  0.00217700
+   10.00000000    0.00000000    0.00000000
+    0.00000000   10.00000000    0.00000000
+    0.00000000    0.00000000   10.00000000
