+++ tests/parsers/fixtures/cp/6.6_cgsteps/aiida.eig
@@ -1,0 +1,9 @@
+  STEP:     10       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+    0.00    0.00    0.00    0.00
+  STEP:     20       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+    0.00    0.00    0.00    0.00
+  STEP:     30       0.00
+   Eigenvalues (eV), kp =   1 , spin =  1
+    0.00    0.00    0.00    0.00
