+++ tests/parsers/fixtures/cp/6.6_cgsteps/aiida.nos
@@ -1,0 +1,3 @@
+     10  7.256653E-04      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     20  1.451331E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     30  2.176996E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
