+++ tests/parsers/fixtures/cp/6.6_cgsteps/aiida.pos
@@ -1,0 +1,12 @@
+     10  0.00072567
+     0.55526004082412E+00     0.91103562191525E+00     0.98872364107694E+00
+    -0.28694368420611E+00    -0.43546306721422E+00    -0.13864038878369E+01
+    -0.32449998478863E+00    -0.50647775154381E+00     0.41213092167249E+00
+     20  0.00145133
+     0.54633172533491E+00     0.89001431069161E+00     0.10191211689058E+01
+    -0.33465562810515E+00    -0.43479416863214E+00    -0.14025150481694E+01
+    -0.32095996857686E+00    -0.50520572575372E+00     0.41123802370397E+00
+     30  0.00217700
+     0.54901641883215E+00     0.88775297676191E+00     0.10546440033360E+01
+    -0.38110893635143E+00    -0.43184324135395E+00    -0.14221576052694E+01
+    -0.31822443015505E+00    -0.50524882533800E+00     0.41024550637084E+00
