+++ tests/parsers/fixtures/cp/6.6_cgsteps/aiida.vel
@@ -1,0 +1,12 @@
+     10  0.00072567
+    -0.47860507809274E-03    -0.99154611695481E-03     0.92351453973289E-03
+    -0.16044323564588E-02    -0.35762583896111E-05    -0.42197916932168E-03
+     0.13018983965946E-03     0.62195148459015E-04    -0.31345960650718E-04
+     20  0.00145133
+    -0.10469499573991E-03    -0.39006457038596E-03     0.11045274768089E-02
+    -0.15732049379737E-02     0.54383180629452E-04    -0.62521105837807E-03
+     0.10486874585709E-03     0.20980086859780E-04    -0.29957276151925E-04
+     30  0.00217700
+     0.27306635185580E-03     0.22369037886466E-03     0.12497048730804E-02
+    -0.15191022093195E-02     0.14829361885881E-03    -0.65286766402546E-03
+     0.77877241091475E-04    -0.23248999857714E-04    -0.37302325565948E-04
