+++ tests/parsers/fixtures/cp/6.6_verlet/aiida.eig
@@ -1,0 +1,9 @@
+  STEP:     40       0.01
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.00  -12.42   -8.02   -6.31
+  STEP:     50       0.01
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.09  -12.48   -8.08   -6.33
+  STEP:     60       0.01
+   Eigenvalues (eV), kp =   1 , spin =  1
+  -24.25  -12.55   -8.17   -6.38
