+++ tests/parsers/fixtures/cp/6.6_verlet/aiida.evp
@@ -1,0 +1,3 @@
+     40  6.119777E-03  1.979794E-04  0.000000E+00  1.723073E+02        -17.17222750        -17.17222750        -17.16977201        -17.16957403  1.000000E+03      -0.00000
+     50  6.845443E-03  1.501844E-04  0.000000E+00  2.477462E+02        -17.17325078        -17.17325078        -17.16972023        -17.16957005  1.000000E+03      -0.00000
+     60  7.571108E-03  1.787469E-04  0.000000E+00  3.422021E+02        -17.17463062        -17.17463062        -17.16975402        -17.16957527  1.000000E+03      -0.00000
