+++ tests/parsers/fixtures/cp/6.6_verlet/aiida.nos
@@ -1,0 +1,3 @@
+     40  6.119777E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     50  6.845443E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
+     60  7.571108E-03      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000      0.00000000
