+++ tests/parsers/fixtures/cp/6.6_verlet/aiida.pos
@@ -1,0 +1,12 @@
+     40  0.00611978
+     0.61619755176652E+00     0.97624655952817E+00     0.11834751991662E+01
+    -0.58066576848775E+00    -0.33933950060878E+00    -0.14165606720626E+01
+    -0.30995094895495E+00    -0.51656115805742E+00     0.40184374830606E+00
+     50  0.00684544
+     0.60933451387917E+00     0.96750721713826E+00     0.11701262445686E+01
+    -0.60563772880505E+00    -0.30484266124150E+00    -0.13910331238958E+01
+    -0.30796126156716E+00    -0.51817100161851E+00     0.40108258620799E+00
+     60  0.00757111
+     0.59455154338114E+00     0.94823974259138E+00     0.11444339524483E+01
+    -0.62795543071568E+00    -0.26479550933008E+00    -0.13652114936866E+01
+    -0.30564246954162E+00    -0.51946973145379E+00     0.40107450257743E+00
