+++ tests/parsers/fixtures/cp/6.6_verlet/aiida.vel
@@ -1,0 +1,12 @@
+     40  0.00611978
+    -0.83288570118430E-04    -0.10178665602975E-03    -0.22464811431683E-03
+    -0.87663612463370E-03     0.10482213779067E-02     0.79084338734048E-03
+     0.59995293422004E-04    -0.59152170117314E-04    -0.35387204563968E-04
+     50  0.00684544
+    -0.37130931809487E-03    -0.48700694011190E-03    -0.66959878875101E-03
+    -0.78147436064745E-03     0.12421820325596E-02     0.88032368666099E-03
+     0.72048979921398E-04    -0.47198443278001E-04    -0.13170306119373E-04
+     60  0.00757111
+    -0.59408922865524E-03    -0.76924082089529E-03    -0.10227831889684E-02
+    -0.71632244915537E-03     0.14250853886425E-02     0.82400942850896E-03
+     0.81900729863154E-04    -0.40990285484210E-04     0.12423360028734E-04
