+++ tests/parsers/fixtures/cp/6.6_verlet/print_counter
@@ -1,0 +1,2 @@
+LAST SUCCESSFUL PRINTOUT AT STEP:
+   60
