+++ tests/parsers/test_cp/test_cp_default_6_6_cgstep_.yml
@@ -1,0 +1,66 @@
+parameters:
+  beta_real_space: false
+  charge_density: ./charge-density.dat
+  conjugate_gradient: true
+  constraint_mag: 0
+  creator_name: cp
+  creator_version: '6.6'
+  dft_exchange_correlation: null
+  do_magnetization: false
+  do_not_use_time_reversal: null
+  exit_status: 0
+  fermi_energy_units: eV
+  fft_grid:
+  - 60
+  - 60
+  - 60
+  format_name: QEXSD
+  format_version: 20.04.20
+  has_dipole_correction: false
+  has_electric_field: false
+  inversion_symmetry: false
+  k_points:
+  - - 0.0
+    - 0.0
+    - 0.0
+  k_points_units: 1 / angstrom
+  k_points_weights:
+  - 2.0
+  lattice_symmetries: []
+  lda_plus_u_calculation: false
+  lkpoint_dir: false
+  lsda: false
+  magnetization_angle1:
+  - 0.0
+  - 0.0
+  magnetization_angle2:
+  - 0.0
+  - 0.0
+  no_time_rev_operations: null
+  non_colinear_calculation: false
+  number_of_atomic_wfc: 6
+  number_of_bands: 4
+  number_of_bravais_symmetries: null
+  number_of_electrons: 8.0
+  number_of_k_points: 1
+  number_of_spin_components: 1
+  number_of_symmetries: null
+  q_real_space: false
+  rho_cutoff: 4353.821352096
+  rho_cutoff_units: eV
+  smooth_fft_grid:
+  - 60
+  - 60
+  - 60
+  spin_orbit_calculation: false
+  spin_orbit_domag: false
+  starting_magnetization:
+  - 0.0
+  - 0.0
+  symmetries: []
+  symmetries_units: crystal
+  time_reversal_flag: true
+  wall_time: '     12.98s '
+  warnings: []
+  wfc_cutoff: -27.2113834506
+  wfc_cutoff_units: eV
