+++ tests/parsers/test_cp/test_cp_default_default_.yml
@@ -136,7 +136,6 @@ parameters:
   number_of_processors_per_taskgroup: 1
   number_of_species: 1
   number_of_spin_components: 1
-  parser_warnings: []
   pp_check_flag: true
   reciprocal_lattice_vectors:
   - - -0.1851851851851852
@@ -209,6 +208,8 @@ trajectory:
   - 10
   array|enthalpy_plus_kinetic:
   - 10
+  array|forces:
+  - 0
   array|ionic_temperature:
   - 10
   array|positions:
